Crate reth_codecs

Compact codec.

Macros

impl_compact_for_bytes

Implements the Compact trait for fixed size byte array types like B256.

Traits

Compact

Trait that implements the Compact codec.

Attribute Macros

add_arbitrary_tests

To be used for types that implement Arbitrary manually. See derive_arbitrary() for more.

derive_arbitrary

Adds Arbitrary and proptest::Arbitrary imports into scope and derives the struct/enum.

main_codec

This code implements the main codec. If the codec supports it, it will also provide the derive_arbitrary() function, which automatically implements arbitrary traits and roundtrip fuzz tests.

no_codec

use_compact

use_postcard

use_scale

Derive Macros

Compact

CompactZstd