Attribute Macro reth_codecs::derive_arbitrary

#[derive_arbitrary]

Adds Arbitrary and proptest::Arbitrary imports into scope and derives the struct/enum.

If compact or rlp is passed to derive_arbitrary, there will be proptest roundtrip tests generated. An integer value passed will limit the number of proptest cases generated (default: 256).

Examples:

#[derive_arbitrary]: will derive arbitrary with no tests.

#[derive_arbitrary(rlp)]: will derive arbitrary and generate rlp roundtrip proptests.

#[derive_arbitrary(rlp, 10)]: will derive arbitrary and generate rlp roundtrip proptests. Limited to 10 cases.

#[derive_arbitrary(compact, rlp)]. will derive arbitrary and generate rlp and compact roundtrip proptests.