Trait reth_codecs::Compact

pub trait Compact: Sized {
    // Required methods
    fn to_compact<B>(self, buf: &mut B) -> usize
       where B: BufMut + AsMut<[u8]>;
    fn from_compact(buf: &[u8], len: usize) -> (Self, &[u8]);

    // Provided methods
    fn specialized_to_compact<B>(self, buf: &mut B) -> usize
       where B: BufMut + AsMut<[u8]> { ... }
    fn specialized_from_compact(buf: &[u8], len: usize) -> (Self, &[u8]) { ... }
}

Trait that implements the Compact codec.

When deriving the trait for custom structs, be aware of certain limitations/recommendations:

Works best with structs that only have native types (eg. u64, B256, U256).

Fixed array types (B256, Address, Bloom) are not compacted.

Max size of T in Option<T> or Vec<T> shouldn’t exceed 0xffff.

Any Bytes field should be placed last.

Any other type which is not known to the derive module should be placed last in they contain a Bytes field.

The last two points make it easier to decode the data without saving the length on the StructFlags. It will fail compilation if it’s not respected. If they’re alias to known types, add their definitions to get_bit_size() or known_types in generator.rs.

Regarding the specialized_to/from_compact methods: Mainly used as a workaround for not being able to specialize an impl over certain types like Vec<T>/Option<T> where T is a fixed size array like Vec<B256>.

Required Methods

fn to_compact<B>(self, buf: &mut B) -> usize
where B: BufMut + AsMut<[u8]>,

Takes a buffer which can be written to. Ideally, it returns the length written to.

fn from_compact(buf: &[u8], len: usize) -> (Self, &[u8])

Takes a buffer which can be read from. Returns the object and buf with its internal cursor advanced (eg..advance(len)).

len can either be the buf remaining length, or the length of the compacted type.

It will panic, if len is smaller than buf.len().

Provided Methods

fn specialized_to_compact<B>(self, buf: &mut B) -> usize
where B: BufMut + AsMut<[u8]>,

“Optional”: If there’s no good reason to use it, don’t.

fn specialized_from_compact(buf: &[u8], len: usize) -> (Self, &[u8])

“Optional”: If there’s no good reason to use it, don’t.

Object Safety

This trait is not object safe.